Prep to add /etc/cron.* managment
This commit moves what was the init.pp into crontab.pp as this is more in line
with the naming convention to manage hourly, daily, weekly and monthly jobs.

A new init.pp has been created which ensures that the cron service is running.

The crontab.pp has also gained support for purging unmanaged cron entries.