Add support for setting hiera_hash globaly
This commit allows $hiera_hash to be set in the man cron class. If set to true
all child classes will use hiera_hash($foo) style lookups for cron jobs.

Previously this was done on a per class basis. This still can be set per class,
although the setting of $hiera_hash to true in the cron class will take
precedence over individual classes which are set to false (default).