Fixed bugs in jobs created by parameters
Attempting to get the namevar to avoid the user explicitly stating the file path
of the job such as '/etc/cron.daily/job1' were not sucessful. As such the onyl
workable solution was to employ the use of the future parser.

There were also some incorrect atributes such as using 'user' for the files
rather than 'owner'.

This commit also removes the '$defaults' parameter in the *ly.pp classes.