Allows for global purging, also fixed global hashing
This commit adds the ability to purge all cron resources by setting the $pruge
parameter to true in the main cron class.

Also fixed behaviour of when setting $hiera_hash to be true in the main cron
class to actually do what is stated in the README. Also made other updates to
the README regarding purging and the fact there is a service class.