Crontab class no longer requires params
As the init.pp includes the crontab class automatically rather than explictly it
may be the case that the user does not pass any parameters for the $jobs
parameter to the class. This commit adds logic to only create jobs if a hash if
passed, otherwise if anything but undef is recived an error is produced.