Removed *ly classes for depuplication
This classes were largely identical but differed in which interval they were
referencing. This resulted in a large amount of duplication and made it easier
to introduce errors with certain classes not being updated identically.

This commit removes these classes in favor of a single interval class in which
the interval can be passed with the job. This will make maintaining the module
much easier. However as this changes name spacing and the parameters to
interface with the module this is a breaking change. As such the version number
has been bumped to 0.1.0. As the layout of the module is now far more concise
I do not envisage that the name spacing will change in a breaking manner
anywhere in the near future.

The README has been updated to reflect these changes as well as having some
whitespace removed. The crontab class has also had it's in line documentation
updated.